+++ src/languages/en.json
@@ -155,7 +155,7 @@
 
     "downtime-dnd5e.AddCategory": "Add Category",
     "downtime-dnd5e.AddWorldCategory": "Add World Category",
-    "downtime-dnd5e.AddItemCategory": "Add Item Cateogry",
+    "downtime-dnd5e.AddActorCategory": "Add Actor Category",
     "downtime-dnd5e.AddItem": "Add Item",
     "downtime-dnd5e.AddWorldItem": "Add World Downtime",
     "downtime-dnd5e.CreateNewCategory": "Add a new category",
