--- src/styles/downtime-dnd5e.css
@@ -1,3 +1,0 @@
-.downtime-dnd5e-form {
-    overflow: hidden;
-}
